Moved logic for popup from app to context. Updated code which used it from app to use context.
